American Sign Language Detection/Web app/requirements.txt
flask
pandas
Pillow
numpy
tensorflow
